+++ default.c
@@ -655,7 +655,7 @@ sd_TeXstyle(const char *v, int64_t flag)
                         NULL };
   ulong n = GP_DATA->fmt->TeXstyle;
   GEN z = sd_ulong(v,flag,"TeXstyle", &n, 0, 7, msg);
-  GP_DATA->fmt->TeXstyle = n; return z;
+  GP_DATA->fmt->TeXstyle = (int)n; return z;
 }
 
 GEN
@@ -669,7 +669,7 @@ sd_output(const char *v, int64_t flag)
                  "(external prettyprint)", NULL};
   ulong n = GP_DATA->fmt->prettyp;
   GEN z = sd_ulong(v,flag,"output", &n, 0,3,msg);
-  GP_DATA->fmt->prettyp = n;
+  GP_DATA->fmt->prettyp = (int)n;
   GP_DATA->fmt->sp = (n != f_RAW);
   return z;
 }
