+++ rootpol.c
@@ -652,7 +652,7 @@ fft2(GEN W, GEN p, GEN f, int64_t step, int64_t l)
     gel(f,i)    = gadd(f0, f1);
     gel(f,i+l1) = gsub(f0, f1);
   }
-  gerepilecoeffs(av, f, l);
+  gerepilecoeffs(av, f, (int)l);
 }
 
 static void
@@ -707,7 +707,7 @@ fft(GEN W, GEN p, GEN f, int64_t step, int64_t l, int64_t inv)
     gel(f,i+l2) = gsub(f02, f13);
     gel(f,i+l3) = gsub(g02, g13);
   }
-  gerepilecoeffs(av, f, l);
+  gerepilecoeffs(av, f, (int)l);
 }
 
 #define code(t1,t2) ((t1 << 6) | t2)
